Refactoring: remove scan driver, combine into multi driver
add new method for multi driver
