+++ tests/Drivers/MultiDriverTest.php
@@ -247,4 +247,50 @@ class MultiDriverTest extends TestCase
         $result = $driver->key($key)->count();
         $this->assertEquals($expected, $result);
     }
+
+    /**
+     * @test
+     */
+    public function it_should_find_keys_by_scan()
+    {
+        $keyPattern = $this->assembleKey('*');
+        $key1 = $this->assembleKey('abc');
+        $key2 = $this->assembleKey('def');
+        $key3 = $this->assembleKey('ghi');
+        $keys = [
+            $key1,
+            $key2,
+            $key3,
+        ];
+        $expected = $keys;
+        $this->testRedisClient->set($keys[0], 1);
+        $this->testRedisClient->set($keys[1], 2);
+        $this->testRedisClient->set($keys[2], 3);
+
+        $driver = new MultiDriver($this->testRedisClient);
+        $cursor = '0';
+        $actual = [];
+        do {
+            list($cursor, $results) = $driver->key($keyPattern)->scan($cursor, 1);
+            foreach ($results as $key) {
+                $actual[] = $key;
+            }
+        } while ($cursor !== '0');
+        sort($expected);
+        sort($actual);
+        $this->assertEquals($expected, $actual);
+    }
+
+    /**
+     * ??
+     */
+    public function it_should_get_empty_response()
+    {
+        $expected = [];
+        $cursor = '0';
+        $count = 1;
+        $driver = new ScanDriver($this->testRedisClient);
+        list($cursor, $actual) = $driver->exec($cursor, $count);
+        $this->assertEquals($expected, $actual);
+    }
 }
